Supplied missing rate() for level, pattern and FlatPlatform. Created Gap Factories. Generator probability distribution work. Horizontal Obstacles

+++ IPCG/src/Model/Component.java
@@ -9,7 +9,7 @@ public abstract class Component extends GameCollection<Part> {
 
 	@Override
 	public double rate() {
-		float rating = 0;
+		double rating = 0;
 		for (Part element : elements) {
 			rating += element.rate();
 		}
